[17] Add BG image (#18)
This PR adds the background image.

It's has slightly more opacity than the original website and it's 
also not the best CSS — it scales weirdly on mobile — but it's ok 
for this part of the project (copying over the old styles).

Also, there's a small fix on the generated Head of the Info page.